Android 14.0.0 release 29

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCZeZW6wAKCRDorT+BmrEO
# eM7kAJ9imUCbC4Y/hDvcAXa6NaIPosiWnQCdHmGPdiWc9+8PqFkW2PvhuQW3khA=
# =zs0h
# -----END PGP SIGNATURE-----
# gpg: Signature made Tue Mar  5 01:19:07 2024 EET
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Good signature from "The Android Open Source Project <initial-contribution@android.com>" [marginal]
# gpg: initial-contribution@android.com: Verified 2369 signatures in the past
#      2 years.  Encrypted 4 messages in the past 2 years.
# gpg: WARNING: This key is not certified with sufficiently trusted signatures!
# gpg:          It is not certain that the signature belongs to the owner.
# Primary key fingerprint: 4340 D135 70EF 945E 8381  0964 E8AD 3F81 9AB1 0E78

# By Kelvin Zhang (52) and others
# Via Automerger Merge Worker (549) and others
* tag 'android-14.0.0_r29': (74 commits)
  Add support to disable multi-threaded compression at install time
  Fix download progress offset error
  Remove dependency on cacerts_google
  Convert int to string for payload sizes
  Add specific error codes to enable better error handling
  Convert CowOperationType into strongly typed enums
  Match cache size between puffdiff and lz4diff
  Add payload_hash/metadata_hash to update scripts
  Add a sysprop to delay merge by fix duration
  update engine: fix gnu++20 build
  Add error check for GetCurrentSlot
  Allow host cow_converter to change compression param
  Link to libprotobuf dynamically
  Improve atomiticy of update checkpointing
  Use ASSERT_* instead of EXPECT_* to abort test early
  Kill the entire child process group when cancelling postinstall
  Allow on device spl downgrade check to be bypassed
  OTA: update SetBlockDeviceReadOnly
  Revert "Add UID permission check to update_engine"
  Add UID permission check to update_engine
  ...

 Conflicts:
	Android.bp

+++ common/cow_operation_convert.h
@@ -26,8 +26,8 @@
 namespace chromeos_update_engine {
 struct CowOperation {
   enum Type {
-    CowCopy = android::snapshot::kCowCopyOp,
-    CowReplace = android::snapshot::kCowReplaceOp,
+    CowCopy = static_cast<int>(android::snapshot::kCowCopyOp),
+    CowReplace = static_cast<int>(android::snapshot::kCowReplaceOp),
   };
   Type op{};
   uint64_t src_block{};
